+++ esp-hal/CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `gpio::interconnect` types now have a lifetime associated with them (#3302)
 - The `critical-section` implementation is now gated behind the `critical-section-impl` feature (#3293)
 - `Trace` is no longer generic (#3305)
+- Migrate SPI slave driver to newer DMA API (#3326)
 
 ### Fixed